% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoad-package.R
\docType{package}
\name{autoad-package}
\alias{autoad}
\alias{autoad-package}
\title{autoad: Meta-Learning for Unsupervised Time Series Anomaly Detection}
\description{
This package contains several model recommendation methods for unsupervised Anomaly Detection time series, alongside utility functions (for data managing) and data (precomputed meta-features and performance matrices).
}
\keyword{internal}
