% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommender_hydra.R
\name{fit.hydra}
\alias{fit.hydra}
\title{Fit Hydra model recommender}
\usage{
fit.hydra(object, train_performance, train_metafeatures, ...)
}
\arguments{
\item{train_performance}{A data frame or matrix corresponding one row per dataset and one column per configuration, with performance values on each cell.}

\item{train_metafeatures}{A a data frame containing the metafeatures of each dataset as columns.}
}
\value{
A list of class `Hydra` containing four elements: `local_head`, a fitted cFact recommender, `global_head`, a fitted URegression recommender, `train_metafeatures`, the
internal data frame used to calculate new datasets distances, and `recommender_type`, a character containing the recommender type ('Hydra').
}
\description{
Bi-headed model that uses UMAP and clustering to detect if new datasets to output recommendations to are close or far from existing datasets in the
meta-feature space. It uses a URegression model for novel datasets and a cFact recommender for existing datasets.
}
\examples{

dat <- get_sample_data(train_test_split = T)
hydra <- recommender('hydra')
hydra <- hydra \%>\% fit(dat$train$performance, dat$train$metafeatures)
hydra \%>\% predict(dat$test$metafeatures)
}
