% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommender_orthos_local.R
\name{fit.orthos_local}
\alias{fit.orthos_local}
\title{Fit a regression model recommender with data clustering}
\usage{
fit.orthos_local(object, train_performance, train_metafeatures, ...)
}
\arguments{
\item{train_performance}{A data frame or matrix corresponding one row per dataset and one column per configuration, with performance values on each cell.}

\item{train_metafeatures}{A a data frame containing the metafeatures of each dataset as columns.}

\item{...}{Extra parameters.}
}
\value{
A list containing two elements: `train_performance` and `train_metafeatures`.
}
\description{
Model recommender training. This model trains Random Forest regression models directly from the metafeatures to the performance matrix, splitting both matrices per row, to separate in data clusters of different behaviors.
}
