% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_precomputed_performance.R
\name{get_precomputed_performance}
\alias{get_precomputed_performance}
\title{Get Precomputed Performances}
\usage{
get_precomputed_performance(
  providers = "all",
  wide = T,
  metric = "affiliation_pr_auc"
)
}
\arguments{
\item{providers}{`all` or any combination of `BGP`, `Benchmark`, `SMD` and `Water`. Defaults to `all`.}

\item{wide}{Logical. Whether to return the performance data frame in wide or long format. Defaults to `TRUE`.}

\item{metric}{One of `all`, `affiliation_pr_auc`, `pr_auc` or `roc`. Which metric to return. Defaults to `affiliation_prauc`. If `all` is selected, `wide` cannot be `TRUE`.}
}
\value{
A data frame of wide format with the selected metric.
}
\description{
This function returns precomputed Precision Recall AUC for a large combination of datasets and configurations. It can be
used to quickly evaluate recommender functions.
}
