% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommender_hydra.R
\name{predict.hydra}
\alias{predict.hydra}
\title{Recommend models for new datasets using Hydra}
\usage{
predict.hydra(object, test_metafeatures)
}
\arguments{
\item{object}{Object obtained from `fit.hydra()`.}

\item{test_metafeatures}{A a data frame containing the metafeatures of each dataset as columns.}
}
\value{
A character string corresponding to the recommended model per row on `test_metafeatures`.
}
\description{
Model recommendation. Issues a single model recommendation per row in `test_metafeatures`.
}
\examples{

dat <- get_data(train_test_split = T)
hydra <- recommender('hydra')
hydra <- hydra \%>\% fit(dat$train$performance, dat$train$metafeatures)
hydra \%>\% predict(dat$test$metafeatures)
}
