% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommender_orthos_local.R
\name{predict.orthos_local}
\alias{predict.orthos_local}
\title{Recommend models for new datasets using regression}
\usage{
predict.orthos_local(object, test_metafeatures)
}
\arguments{
\item{object}{Object obtained from `fit.regression_rf()`.}

\item{test_metafeatures}{A a data frame containing the metafeatures of each dataset as columns.}
}
\value{
A character string corresponding to the recommended model per row on `test_metafeatures`.
}
\description{
Model recommendation. Issues a single model recommendation per row in `test_metafeatures`.
}
